+++ src/lib/util/log/types/LoggerConfig.ts
@@ -1,8 +1,0 @@
-import { LogTransport } from './'
-
-export interface LoggerConfig {
-	enable: boolean
-	transports: LogTransport[]
-	default?: boolean
-	name?: string
-}