+++ src/lib/util/schedule/Scheduler.ts
@@ -2,7 +2,8 @@ import { Job } from 'node-schedule'
 import { log } from '~/lib/util'
 import { AbstractTask, AbstractTaskGroup } from './'
 
-export class Scheduler {
+export class Scheduler
+{
 	/*
 	 * Runnable Job/Task(s) Stack
 	 */
@@ -16,99 +17,98 @@ export class Scheduler {
 	/*
 	 * Construct
 	 */
-	constructor(stack: (AbstractTask | AbstractTaskGroup)[] = []) {
-		// Set stack
+	constructor(stack: (AbstractTask | AbstractTaskGroup)[] = [])
+	{
 		this.stack = stack
 	}
 
 	/*
 	 * Start Running Schedule
 	 */
-	public async start(): Promise<void> {
-		// Log info
-		log.info('Schedule: starting')
+	public async start(): Promise<void>
+	{
+		log.info('Schedule starting')
 
 		// Check if already running
-		if (this.job.length > 0) {
-			// Log error
-			log.error('Schedule: unable to start schedule, jobs running')
+		if(this.job.length > 0)
+		{
+			log.error('Schedule unable to start, jobs already running')
 			return
 		}
 
 		// Iterate runnable stack
-		for (let i = 0; i < this.stack.length; i++) {
-			// Get task/job
+		for(let i = 0; i < this.stack.length; i++)
+		{
 			const job = this.stack[i]
 
 			// Create cron job
-			this.job[i] = new Job(job.getName(), () => {
+			this.job[i] = new Job(job.getName(), () =>
+			{
 				job.run()
 			})
 
 			// Schedule job
 			const scheduled = this.job[i].schedule(job.getSchedule())
 
-			// Success
-			if (scheduled) {
-				// Log info
-				log.info('Schedule: job scheduled', {
-					name: job.getName(),
-					schedule: job.getSchedule(),
-					description: job.getDescription()
-				})
+			if(scheduled)
+			{
+				log.info(
+					{ name: job.getName(), schedule: job.getSchedule(), description: job.getDescription() },
+					'Schedule job scheduled'
+				)
 			}
 		}
 
-		// All jobs started
-		if (this.job.length > 0 && this.stack.length === this.job.length) {
-			// Log info
-			log.info('Schedule: started')
+		if(this.job.length > 0 && this.stack.length === this.job.length)
+		{
+			log.info('Schedule started')
 		}
 	}
 
 	/*
 	 * Stop Running Schedule
 	 */
-	public async stop(): Promise<void> {
-		// Log info
-		log.info('Schedule: stopping')
+	public async stop(): Promise<void>
+	{
+		log.info('Schedule stopping')
 
-		// Iterate runnable stack
-		for (let i = 0; i < this.job.length; i++) {
-			// Cancel job
+		for(let i = 0; i < this.job.length; i++)
+		{
 			const cancelled = this.job[i].cancel()
 
-			// Cancelled?
-			if (cancelled) {
-				// Remove job
+			if(cancelled)
+			{
 				this.job.splice(i, 1)
-			} else {
-				// Log error
-				log.error('Schedule: job not cancelled', this.stack[i])
+			}
+			else
+			{
+				log.error({ task: this.stack[i]?.getName() }, 'Schedule job not cancelled')
 			}
 		}
 
-		// Log based on results
-		if (this.job.length > 0) {
-			// Log error
-			log.error('Schedule: jobs not stopped', { job: this.job })
-		} else {
-			// Log info
-			log.info('Schedule: stopped')
+		if(this.job.length > 0)
+		{
+			log.error({ remaining: this.job.length }, 'Schedule jobs not stopped')
+		}
+		else
+		{
+			log.info('Schedule stopped')
 		}
 	}
 
 	/*
 	 * Job property getter
 	 */
-	public getJob(): Job[] {
+	public getJob(): Job[]
+	{
 		return this.job
 	}
 
 	/*
 	 * Stack property getter
 	 */
-	public getStack(): (AbstractTask | AbstractTaskGroup)[] {
+	public getStack(): (AbstractTask | AbstractTaskGroup)[]
+	{
 		return this.stack
 	}
 }