+++ src/lib/util/storage/drivers/index.ts
@@ -1,0 +1,2 @@
+export { localDriver } from './local'
+export { s3Driver } from './s3'